int changed with long

+++ APBDFinalProject/APBDFinalProject/Repositories/Abstractions/IBusinessCustomerRepository.cs
@@ -5,6 +5,6 @@ namespace APBDFinalProject.Repositories;
 public interface IBusinessCustomerRepository
 {
     Task<int> AddCustomer(BusinessCustomer businessCustomer, CancellationToken cancellationToken);
-    Task<int> UpdateCustomer(int krs, BusinessCustomer businessCustomer, CancellationToken cancellationToken);
-    Task<BusinessCustomer?> GetBusinessCustomerByKRS(int krs, CancellationToken cancellationToken);
+    Task<int> UpdateCustomer(long krs, BusinessCustomer businessCustomer, CancellationToken cancellationToken);
+    Task<BusinessCustomer?> GetBusinessCustomerByKRS(long krs, CancellationToken cancellationToken);
 }
